refactor: Change rating_user from float to int on movie and tv entity

+++ consumatio/entities/movie.py
@@ -20,7 +20,7 @@ class Movie(Entity):
     directors: list
     tmdb_url: str
     watch_status: str
-    rating_user: float
+    rating_user: int
     favorite: bool
 
     @classmethod
